Google calendar api integrated and integration related things are managed in table

+++ builddb.php
@@ -1,12 +1,13 @@
 <?php
-include ("connection/mysqlconnection.php");
-$dropdb = "DROP DATABASE $dbName";
-$sql = "CREATE DATABASE $dbName";
-if ($conn->query($dropdb) === TRUE && $conn->query($sql) === TRUE) {
-    echo "Database created successfully\n";
-    //Create require table once database is created
-    $tableQuery = "create table events(ID int primary key,event_name text,event_description TEXT,location TEXT,date TEXT);";
-    $tableQuery = "
+include 'connection/mysqlconnection.php';
+
+if ($conn->ping()) {
+    $dropdb = "DROP DATABASE $dbName";
+    $sql = "CREATE DATABASE $dbName";
+    if ($conn->query($dropdb) === TRUE && $conn->query($sql) === TRUE) {
+        echo "Database created successfully\n";
+        //Create require table once database is created
+        $eventsTableQuery = "
     CREATE TABLE `events` (
         `id` int(11) NOT NULL AUTO_INCREMENT,
         `name` varchar(255) COLLATE utf8_unicode_ci NOT NULL,
@@ -18,13 +19,28 @@ if ($conn->query($dropdb) === TRUE && $conn->query($sql) === TRUE) {
         `google_calendar_event_id` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
         `created` datetime NOT NULL DEFAULT current_timestamp(),
         PRIMARY KEY (`id`)
-      ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
-    ";
-    $conn->select_db($dbName);
-    if ($conn->query($tableQuery) === TRUE) {
-        echo "Event table created successfully\n\n";
+      ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;";
+        $conn->select_db($dbName);
+        if ($conn->query($eventsTableQuery) === TRUE) {
+            echo "Event table created successfully\n";
+            $integrationTableQuery = "
+        CREATE TABLE `integrations` (
+        `id` int(11) NOT NULL AUTO_INCREMENT,
+        `data`  text COLLATE utf8_unicode_ci NOT NULL,
+        `status` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
+        `service` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
+        `token` text COLLATE utf8_unicode_ci NOT NULL,
+        `created` datetime NOT NULL DEFAULT current_timestamp(),
+        PRIMARY KEY (`id`)
+      ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;";
+            if ($conn->query($integrationTableQuery) === TRUE) {
+                echo "Integration table created successfully\n\n";
+            }
+        }
+    } else {
+        echo "Error creating database: " . $conn->error;
     }
 } else {
-    echo "Error creating database: " . $conn->error;
+    echo "Error: " . $conn->error;
 }
 $conn->close();
